JMXFetch cycle management on Windows
