+++ pkg/collector/corechecks/embed/jmx/runner.go
@@ -8,7 +8,6 @@
 package jmx
 
 import (
-	"runtime"
 	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
@@ -30,10 +29,6 @@ func (r *runner) initRunner() {
 func (r *runner) startRunner() error {
 
 	lifecycleMgmt := true
-	if runtime.GOOS == "windows" {
-		lifecycleMgmt = false
-	}
-
 	err := r.jmxfetch.Start(lifecycleMgmt)
 	if err != nil {
 		s := status.JMXStartupError{LastError: err.Error(), Timestamp: time.Now().Unix()}
